tmail: cast to size_t in memset()

+++ src/sys/gethostname.c
@@ -26,7 +26,7 @@ char *hostname(void)
 {
 	char hostname[HOST_NAME_MAX];
 
-	memset(hostname, 0, HOST_NAME_MAX);
+	memset(hostname, 0, (size_t)HOST_NAME_MAX);
 
 	if (gethostname(hostname, HOST_NAME_MAX) != 0)
 		return NULL;
